Sneaky bugfixes.

+++ code/game/verbs/suicide.dm
@@ -181,6 +181,11 @@
 	set category = "pAI Commands"
 	set desc = "Kill yourself and become a ghost (You will receive a confirmation prompt)"
 	set name = "pAI Suicide"
+
+	if(!istype(loc,/obj/item/device/paicard))
+		src << "You must be back in your immobile form to wipe your core files."
+		return
+
 	var/answer = input("REALLY kill yourself? This action can't be undone.", "Suicide", "No") in list ("Yes", "No")
 	if(answer == "Yes")
 		var/obj/item/device/paicard/card = loc
